requirements.txt
streamlit==1.32.2
torch==2.2.1
torchvision==0.17.1
timm==0.9.12
pillow==10.2.0
